bot/requirements.txt
python-telegram-bot==22.1
requests==2.32.3
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10 